refactor(sections): extract entry-line splitting into entry.rs
Pure motion: the gap finders (column, sentence, =, :, dash, placeholder
boundary), the separator strippers, and the option-list placeholder
vocabulary, with their tests.
